feat: readable streams
