+++ tomb-wasm/Cargo.toml
@@ -38,6 +38,7 @@ web-sys = { version = "0.3.64", features = [
     "RequestInit",
     "RequestMode",
     "Response",
+    "ReadableStream",
     "Window",
 ]}
 wee_alloc = { version = "0.4.5", optional = true }
@@ -45,6 +46,7 @@ wnfs-wasm = "0.1.20"
 wnfs-common = "0.1.20"
 wasm-bindgen = "0.2.86"
 js-sys = "0.3.64"
+futures-util = "0.3.28"
 
 [dev-dependencies]
 wasm-bindgen-test = "0.3.36"
